SOIL updated, adding UserTexture
git-svn-id: https://projectm.svn.sourceforge.net/svnroot/projectm/trunk@1108 6778bc44-b910-0410-a7a0-be141de4315d

+++ src/projectM-engine/image_DXT.c
@@ -533,7 +533,7 @@ void
 	vec_len2 = 0.0f;
 	for( i = 0; i < 3; ++i )
 	{
-		color_line[i] = c1[i] - c0[i];
+		color_line[i] = (float)(c1[i] - c0[i]);
 		vec_len2 += color_line[i] * color_line[i];
 	}
 	if( vec_len2 > 0.0f )
